[REFACTOR] Extract SDK to its own module
